# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ASOTRACMET · Enturnamiento",
    "version": "1.0.0",
    "description": "API REST del sistema de enturnamiento (spec §8). Errores con código estable (`ErrorApi`); sesiones Bearer opacas y revocables; RBAC por recurso (`x-guard`); `x-reauth` exige re-autenticación reciente."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/openapi.json": {
      "get": {
        "tags": ["Salud"],
        "summary": "Este contrato (OpenAPI 3.1)",
        "operationId": "getOpenapiJson",
        "x-guard": "público",
        "security": [],
        "responses": {
          "200": {
            "description": "Documento OpenAPI 3.1",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "Documento OpenAPI 3.1",
                  "x-tipo": "OpenApi"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/ofertas/{id}/aceptar": {
      "post": {
        "tags": ["Operación"],
        "summary": "Aceptar (member, solo la propia): genera el TR",
        "operationId": "postOfertasIdAceptar",
        "x-guard": "ofertas A (own)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aceptacion",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Aceptacion"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorApi": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "CHALLENGE_INVALIDO",
              "CODIGO_INVALIDO",
              "FORBIDDEN",
              "FORBIDDEN_OWN_SCOPE",
              "REAUTH_REQUERIDA",
              "NOT_FOUND",
              "TOTP_NO_ENROLADO",
              "COLA_LOCKED",
              "EMAIL_EN_USO",
              "PLACA_EN_USO",
              "DOCUMENTO_EN_USO",
              "CATALOGO_EN_USO",
              "TARIFA_DUPLICADA",
              "VEHICULO_CON_HISTORIA",
              "TR_DUPLICADO",
              "OFERTA_NO_ABIERTA",
              "REQUERIMIENTO_CERRADO",
              "REQUERIMIENTO_SIN_CUPOS",
              "TR_NO_CANCELABLE",
              "VERSION_CONFLICT",
              "VIAJE_YA_EXISTE",
              "VIAJE_LIQUIDADO",
              "VIAJE_ANULADO",
              "RECAUDO_CERRADO",
              "COLA_VACIA",
              "VEHICULO_NO_HABILITADO",
              "VEHICULO_NO_ACTIVO",
              "DOCUMENTO_VENCIDO",
              "BLOQUEO_TEMPORAL",
              "OFERTA_ABIERTA_PREVIA",
              "TR_ACTIVO",
              "PENALIZACION_PENDIENTE",
              "OFERTA_EXPIRADA",
              "MOTIVO_REQUERIDO",
              "TR_NO_VIAJABLE",
              "VIAJE_NO_LIQUIDABLE",
              "PAGO_INVALIDO",
              "RATE_LIMITED",
              "DEMASIADOS_INTENTOS",
              "INVARIANTE_COLA",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": ["code", "message"],
        "additionalProperties": false
      },
...
      "Aceptacion": {
        "type": "object",
        "properties": {
          "oferta": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "requerimientoId": {
                "type": "string"
              },
              "vehiculoId": {
                "type": "string"
              },
              "placa": {
                "type": ["string", "null"]
              },
              "claseCola": {
                "anyOf": [
                  {
                    "type": "string",
                    "enum": ["C100", "C350", "C600", "MM", "TM-CBZ"]
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "etiqueta": {
                "type": ["string", "null"]
              },
              "estado": {
                "type": "string",
                "enum": ["abierta", "aceptada", "declinada", "expirada", "anulada"]
              },
              "ofrecidaEn": {
                "type": "string"
              },
              "expiraEn": {
                "type": "string"
              },
              "motivoDeclinacion": {
                "type": ["string", "null"]
              },
              "nota": {
                "type": ["string", "null"]
              },
              "requerimiento": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "clienteId": {
                        "type": "string"
                      },
                      "cliente": {
                        "type": ["string", "null"]
                      },
                      "clienteNombre": {
                        "type": ["string", "null"]
                      },
                      "destino": {
                        "type": ["string", "null"]
                      },
                      "claseCola": {
                        "type": "string",
                        "enum": ["C100", "C350", "C600", "MM", "TM-CBZ"]
                      },
                      "fechaServicio": {
                        "type": "string"
                      },
                      "cantidadCupos": {
                        "type": "number"
                      },
                      "cuposAsignados": {
                        "type": "number"
                      },
                      "ofertasAbiertas": {
                        "type": "number"
                      },
                      "cuposDisponibles": {
                        "type": "number"
                      },
                      "observaciones": {
                        "type": ["string", "null"]
                      },
                      "estado": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "clienteId",
                      "cliente",
                      "clienteNombre",
                      "destino",
                      "claseCola",
                      "fechaServicio",
                      "cantidadCupos",
                      "cuposAsignados",
                      "ofertasAbiertas",
                      "cuposDisponibles",
                      "observaciones",
                      "estado"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "id",
              "requerimientoId",
              "vehiculoId",
              "placa",
              "claseCola",
              "etiqueta",
              "estado",
              "ofrecidaEn",
              "expiraEn",
              "motivoDeclinacion",
              "nota",
              "requerimiento"
            ],
            "additionalProperties": false
          },
          "tr": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "codigo": {
                "type": "string"
              },
              "placa": {
                "type": ["string", "null"]
              },
              "cliente": {
                "type": ["string", "null"]
              },
              "destino": {
                "type": ["string", "null"]
              },
              "etiqueta": {
                "type": ["string", "null"]
              },
              "fechaAsignacion": {
                "type": "string"
              },
              "estado": {
                "type": "string",
                "enum": ["asignado", "en_curso", "cumplido", "cancelado", "no_tramitar"]
              },
              "motivoCancelacion": {
                "type": ["string", "null"]
              }
            },
            "required": [
              "id",
              "codigo",
              "placa",
              "cliente",
              "destino",
              "etiqueta",
              "fechaAsignacion",
              "estado",
              "motivoCancelacion"
            ],
            "additionalProperties": false
          }
        },
        "required": ["oferta", "tr"],
        "additionalProperties": false
      },
...
    }
  }
}